fix: move providers back to app-level
