+++ frontend/src/app/providers/solana-provider.tsx
