+++ frontend/src/pages/accounts/list/list.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react'
-import { WalletButton } from '@/features/solana-provider'
+import { WalletButton } from '@/app/providers/solana-provider'
 
 export const List: FC = () => (
   <div className="hero py-[64px]">
