+++ frontend/src/shared/ui/header/header.tsx
@@ -6,9 +6,9 @@ import { useWallet } from '@solana/wallet-adapter-react'
 import { useWalletModal } from '@solana/wallet-adapter-react-ui'
 import { cva } from 'class-variance-authority'
 import * as Icons from 'lucide-react'
+import { WalletButton } from '@/app/providers/solana-provider'
 import { ClusterSelect } from '@/entities/cluster/cluster'
 import { DevModeButton } from '@/entities/dev-mode'
-import { WalletButton } from '@/features/solana-provider'
 import { cn } from '@/shared/utils'
 import { Logo } from './logo'
